Moved AdjacencyMatrixTransitionView to other transitions views
