+++ src/main/kotlin/automaton/constructor/view/elements/transition/AdjacencyMatrixTransitionView.kt
@@ -1,4 +1,4 @@
-package automaton.constructor.view.elements.vertex
+package automaton.constructor.view.elements.transition
 
 import automaton.constructor.model.element.Transition
 import automaton.constructor.view.elements.transition.TableTransitionView
